requirements.txt
discord.py
aiosqlite
youtube-dl
